manager/python/pyproject.toml
[tool.poetry]
package-mode = false
authors = ["Miko <jherjati@gmail.com>"]
description = ""
name = "braga-py-worker"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
dramatiq = "^1.14.2"
dramatiq-pg = "^0.11.0"
laspy = {extras = ["laszip", "lazrs"], version = "^2.5.3"}
minio = "^7.1.17"
psycopg2-binary = "^2.9.8"
py3dtiles = {extras = ["las"], version = "^9.0.0"}
python = ">=3.10,<3.13"
python-dotenv = "^1.0.0"
lxml = "^6.0.0"
pystac-client = "^0.9.0"
planetary-computer = "^1.0.0"

[tool.poetry.group.dev.dependencies]
watchdog = "^3.0.0"
gdal = "^3.11.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
